- security bugfix: plain text passwords aren't stored in session - security bugfix: session is excluded from database backup

+++ lib/LMS.class.php
@@ -156,6 +156,10 @@ class LMS
 		{
 			foreach($this->DB->ListTables() as $tablename)
 			{
+			        // skip sessions table for security
+				if($tablename == 'sessions')
+				        continue;
+
 				fputs($dumpfile,"DELETE FROM $tablename;\n");
 				$this->DB->Execute('SELECT * FROM '.$tablename);
 				while($row = $this->DB->_driver_fetchrow_assoc())
